% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cite_RNAseqGSEA.R
\name{cite_RNAseqGSEA}
\alias{cite_RNAseqGSEA}
\title{RNAseq GSEA methods citation}
\usage{
cite_RNAseqGSEA(
  fold_changes = 1.5,
  normalisation_type = "LRT_ruvR_upperquartile"
)
}
\arguments{
\item{fold_changes}{[type: vector<numeric>, default: 1.5] A vector of numerics, get concatenated and formatted for printing.}

\item{normalisation_type}{[type: character, default: "LRT_ruvR_upperquartile"] Either "LRT_ruvR_upperquartile" or "TMM", depends on if replicates could be used for analysis and normalisation type.}
}
\description{
Prints methods used for RNAseq and GSEA analysis, allows for variable interpolation to print a custom message.
}
\examples{
cite_RNAseqGSEA(fold_changes = c(1.5, 2.0), normalisation_type = "TMM")


}
