% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combatCorrected.R
\docType{class}
\name{combatRNAseq-class}
\alias{combatRNAseq-class}
\alias{combatRNAseq}
\title{\code{S4} class; does sva::ComBat_seq batch correction}
\value{

}
\description{
\code{S4} class; does sva::ComBat_seq batch correction
}
\section{Slots}{

\describe{
\item{\code{corrected}}{data.table.}

\item{\code{batches}}{list.}

\item{\code{batch_map}}{data.table.}

\item{\code{qc_plots}}{list.}
}}

\author{
Dereck de Mezquita

This class with its constructor makes it easy to do batch correction. Simply provide the list of \code{data.table}s to the class instantiator. This will return the corrected dataset along with PCA plots of before and after in a list.

The \code{data.table}s must have the columns:
\enumerate{
\item \code{GeneID}
\item \code{GeneBiotype}
\item \code{GeneSymbol}
}
}
