% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtcolnames.R
\name{dtcolnames}
\alias{dtcolnames}
\title{Get the data column names from data.table; exclude ID col}
\usage{
dtcolnames(object, excluded = "sample")
}
\arguments{
\item{object}{[type: data.table] A data.table with an identifier column or more.}

\item{excluded}{[type: vector<character>, default: "sample"] A character vector of one or more column names to exclude.}
}
\value{
character<vector>
}
\description{
Get the data column names from data.table; exclude ID col
}
