% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excelList.R
\name{excelList}
\alias{excelList}
\title{Read an excel workbook to a named list of data.tables or other type}
\usage{
excelList(path, FUN_type = "data.table::data.table", ...)
}
\arguments{
\item{path}{[type: char] a path to an excel file, read by readxl::excel_sheets.}

\item{FUN_type}{[type: function | char, default: "data.table::data.table"] a function or string matching a function name to which to format the resulting data in the list.}

\item{...}{extra arguments passed to readxl::read_excel().}
}
\value{
list<data.table | dx>
}
\description{
You can tell the function what type you want the data.frames to be coerced to. Default is data.table.
}
\examples{

excelList(system.file("extdata", "comparisons.xlsx", package = "derecksLabTools"), FUN_type = as.data.frame)

excelList(system.file("extdata", "comparisons.xlsx", package = "derecksLabTools"), FUN_type = "data.table::as.data.table")

}
