% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list2Excel.R
\name{list2Excel}
\alias{list2Excel}
\title{Save a named list of \code{data.frames} | \code{data.tables} to and Excel workbook with multiple sheets}
\usage{
list2Excel(list, file, ...)
}
\arguments{
\item{list}{[type: list<data.frame | data.table>] A named list of \code{data.frame} or \code{data.table}; note the names should not exceed Excels max char limit of 31.}

\item{file}{[type: character] A file to save to.}

\item{...}{Other arguments passed to: \code{openxlsx::saveWorkbook}.}
}
\description{
Save a named list of \code{data.frames} | \code{data.tables} to and Excel workbook with multiple sheets
}
