% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Volcano.R
\name{plot,Volcano-method}
\alias{plot,Volcano-method}
\title{\code{S4} method for \code{Volcano} \code{S4} class}
\usage{
\S4method{plot}{Volcano}(x = object, title = "", subtitle = "", label = TRUE, ...)
}
\arguments{
\item{Volcano}{}
}
\value{
gg ggplot
}
\description{
Prepare a \code{Volcano} class using the constructor, be sure to set your desired parameters. Then plot.
}
