% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.gmt.R
\name{read.gmt}
\alias{read.gmt}
\title{Read a .gmt pathway type file as a list.}
\usage{
read.gmt(file, remove_description = FALSE)
}
\arguments{
\item{file}{[type: character] Path to the file.}

\item{remove_description}{[type: logical, default: FALSE] Whether to remove description (second element of pathway; first is pathway name) of the pathway.}
}
\value{
list<character>
}
\description{
Read a .gmt pathway type file as a list.
}
\examples{

}
