% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig2UpDownGmt.R
\name{sig2UpDownGmt}
\alias{sig2UpDownGmt}
\title{Signature(s) files to GMT list; split up and down (inclusive) based on log2fc}
\usage{
sig2UpDownGmt(
  signatures_dir,
  pattern = NULL,
  gsub_cols_regex = "EdgeR\\\\.upperquartile_LRT_RUVr_",
  to_upper = TRUE
)
}
\arguments{
\item{signatures_dir}{[type: character] A directory containing signature files.}

\item{pattern}{[type: character<regex>, default: NULL] A regular expression to match files on read-in.}

\item{gsub_cols_regex}{[type: character, default: "EdgeR\\.upperquartile_LRT_RUVr_"] A regular expression to clean comparison names and thus pathway names.}

\item{to_upper}{[type: logical, default: TRUE] Convert genes to uppercase; typically yes, lowercase genes are mice type - GSEA works with uppercase humanised genes.}
}
\value{
list<character>
}
\description{
Signature(s) files to GMT list; split up and down (inclusive) based on log2fc
}
\examples{

}
