% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabs2table.R
\name{tabs2table}
\alias{tabs2table}
\title{Combine Excel sheets to single table}
\usage{
tabs2table(
  path,
  output = "compiled-comparisons.xlsx",
  columns = 3,
  return = FALSE,
  ...
)
}
\arguments{
\item{path}{[type: character] A path to a directory; this will read all \code{.xlsx} files at this location.}

\item{output}{[type: character, default: "compiled-comparisons.xlsx"] The name of the output file - must have extension \code{.xlsx}.}

\item{columns}{[type: numeric, default: 3] Integer defines the number of columns to split the combined tables over. This splits the data and thus avoids having to scroll over a large amount of tables.}

\item{return}{[type: logical, default: FALSE] Boolean if TRUE returns the parsed data.}

\item{...}{Extra arguments to pass to \link[openxlsx]{\code{saveWorkbook}}}
}
\value{
Returns if return arguemnt set to TRUE; a list of \code{data.frame}s - might be useful for analysis - the primary output is the file output.
}
\description{
Combine all sheets (tabs) from one or more Excel workbooks to a single table (an index is generated - first tab), padding is added (empty rows and columns) between the indvidual tables. This is useful for getting an overview of your data and avoiding having to click n tabs.
}
\examples{
dir.create("./test")
file.copy(system.file("extdata", "comparisons.xlsx", package = "derecksLabTools"), "./test")
tabs2table(
    path = "./test",
    output = "output-file.xlsx",
    columns = 3,
    return = FALSE,
    overwrite = TRUE
)

tabs2table(
    path = "./test/comparisons.xlsx",
    output = "output-file.xlsx",
    columns = 3,
    return = FALSE,
    overwrite = TRUE
)

}
\keyword{tabs2table}
