% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to.data.frame.R
\name{to.data.frame}
\alias{to.data.frame}
\alias{to.data.frame.data.table}
\title{\code{S3} generic; convert to data.frame \strong{move} column to \code{rownames}}
\usage{
to.data.frame(x, ...)

\method{to.data.frame}{data.table}(x, id.col = NULL, drop.id.col = TRUE)
}
\arguments{
\item{x}{\link{type: data.table} A \code{data.table} object with an identifier column.}

\item{...}{}
}
\value{
data.table
}
\description{
Useful when working with \code{data.table}s; cannot have \code{rownames}. Other methods (\code{as.data.frame}, \code{data.table::setDF}) allow to convert to a \code{data.frame} but then another step is necessary to move an identifier column to \code{rownames}. This function allows the two in one step.
}
\author{
Dereck de Mezquita
}
