% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valueCoordinates.R
\name{valueCoordinates}
\alias{valueCoordinates}
\title{Coordinates of X values in a data.table|data.frame}
\usage{
valueCoordinates(obj, value = NA)
}
\arguments{
\item{value}{[type: ANY, default: NA]}

\item{dt}{[type: data.table | data.frame]}
}
\value{
data.frame
}
\description{
Sometimes knowing NAs or X value are present in your data is not enough, you want to know where exactly.
}
\details{
This function does:

data == value | is.na(data)

To create a truth table and then retrieves the column and row of where the value occurred as a data.frame.
}
\examples{
test <- head(iris, 10)
test[3, 1] <- NA

valueCoordinates(test, value = NA)

}
