% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.gmt.R
\name{write.gmt}
\alias{write.gmt}
\title{Write GMT list to a formatted file.}
\usage{
write.gmt(list_gmt, output_file)
}
\arguments{
\item{list_gmt}{\link[=type: list<character>]{type: list<character>} List of GMT pathways.}

\item{output_file}{\link{type: character} Output file name.}
}
\description{
Adds new lines per pathway and "\verb{\\t}" as a separator.
}
\examples{

}
