% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2tabs.R
\name{table2tabs}
\alias{table2tabs}
\title{Parse Excel tables from one sheet to named tabs}
\usage{
table2tabs(file, output, table_id = "ID", return = FALSE, ...)
}
\arguments{
\item{file}{[type: character] Path to a file type xlsx.}

\item{output}{[type: character] The name of the output file - must have extension \code{.xlsx}.}

\item{table_id}{[type: character, default: "ID"] This is used for identifying the individual tables on a single sheet.}

\item{return}{[type: logical, default: FALSE] If TRUE returns the parsed data.}

\item{...}{Extra arguments to pass to \link[openxlsx]{\code{saveWorkbook}}}
}
\value{
Returns if return argument set to TRUE; a list of \code{data.frame}s - might be useful for analysis - the primary output is the file output.
}
\description{
Parses tables from one Excel sheet based on an identifier, empty rows/columns must be left between tables as these are used for edge detection by the \code{is.na()} function. Each table on the sheet should have column"" names, the first is used for identification of tables, the second for tab names.
This is a tool used at our lab for quickly writing comparisons for the RNAseq analysis and then converting them to multiple tabs.
The typical format is; colnames: "ID", "comparison_name", where ID designates the sample ID's and comparison_name designates test/control.
}
\details{
Note that you can have other content on your excel sheet as long as it does not contain the table_id string used for parsing.
}
\examples{
path <- system.file("extdata", "comparisons-setup.xlsx", package = "derecksLabTools")
table2tabs(
    file = path,
    output = "output-file.xlsx",
    table_id = "ID",
    return = FALSE,
    overwrite = TRUE
)

}
\keyword{table2tabs}
