% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabs2table.R
\name{tabs2table}
\alias{tabs2table}
\title{Combine Excel sheets to single table}
\usage{
tabs2table(dir = "", columns = 3, out_file = "", return = FALSE)
}
\arguments{
\item{dir}{String; path to a directory; this will read all \code{.xlsx} files at this location.}

\item{columns}{Integer [default 3]; defines the number of columns to split the combined tables over. This splits the data and thus avoids having to scroll over a large amount of tables.}

\item{out_file}{String; the name of the output file - must have extension \code{.xlsx}.}

\item{return}{Boolean [default FALSE]; if TRUE returns the parsed data.}
}
\value{
Returns if return arguemnt set to TRUE; a list of \code{data.frame}s - might be useful for analysis - the primary output is the file output.
}
\description{
Combine all sheets (tabs) from one or more Excel workbooks to a single table (an index is generated - first tab), padding is added (empty rows and columns) between the indvidual tables. This is useful for getting an overview of your data and avoiding having to click n tabs.
}
\keyword{tabs2table}
